AdvancedStorage: UX changes for simplified experience
Instead of hiding this feature in the 3 Dot menu, let's
make it it's own action with a custom icon indicating the
user that it exist.

USB Unlocked Data Access is cool, but appeals to the enthusiasts.
Let's make it more simple to a common user by rebranding this to
USB Connection.

Change-Id: Iaeed8e1ce349b7a3bb92e3fa5a94bf83185f270a